Read speciesLink datasets

Returns a tibble with all or filtered datasets available in the speciesLink for searching.

splinkr_datasets(filter = NULL)

Arguments

filter
Character vector containing one or more terms to filter datasets

Value

data.frame() with the datasets

Examples

# return all datasets
splinkr_datasets()

#> # A tibble: 529 x 9
#>    collectionCode collectionName institutionCode institutionName scope
#>    <chr>          <chr>          <chr>           <chr>           <chr>
#>  1 AcariESALQ     Coleção de Ác… USP             Universidade d… anim…
#>  2 ALCB           Herbário Alex… UFBA            Universidade F… plan…
#>  3 AMNH-Bee       Collaborative… AMNH            American Museu… anim…
#>  4 ANSP-Ichthyol… The Academy o… ANSP            The Academy of… anim…
#>  5 Aranhas-Solob… Coleção de Ar… UFPR            Universidade F… anim…
#>  6 ARAR           Coleção Aracn… UPF             Universidade d… anim…
#>  7 ASE            Herbário da U… UFS             Universidade F… plan…
#>  8 ASU-Lichen     Arizona State… ASU             Arizona State … plan…
#>  9 ASU-Plants     Arizona State… ASU             Arizona State … plan…
#> 10 B              Herbarium Ber… FU-Berlin       Freie Universi… plan…
#> # … with 519 more rows, and 4 more variables: numberOfRecords <chr>,
#> #   location <chr>, keywords_pt <chr>, keywords_en <chr>

# Filtering datasets. Terms order defines filtering
# The second term will be searched in subset of datasets created
# by the first term. The third term will be searched in the (sub)subset of
# the second term and so on.
splinkr_datasets(filter = c("universidade", "xiloteca", "Brasil"))

#> # A tibble: 17 x 9
#>    collectionCode collectionName institutionCode institutionName scope
#>    <chr>          <chr>          <chr>           <chr>           <chr>
#>  1 BOTUw          "Xiloteca \"P… UNESP           Universidade E… plan…
#>  2 CESJ           "Herbário Leo… UFJF            Universidade F… plan…
#>  3 COR            "Herbário da … UFMS            Universidade F… plan…
#>  4 HDCF           "Herbário do … UFSM            Universidade F… plan…
#>  5 HDJF           "Herbário Den… UFVJM           Universidade F… plan…
#>  6 HDJFw          "Xiloteca do … UFVJM           Universidade F… plan…
#>  7 HERBAM         "Herbário da … UNEMAT          Universidade d… plan…
#>  8 HPUC-MG        "Herbário do … PUC-MINAS       Pontifícia Uni… plan…
#>  9 HUCPE          "Herbário da … UNICAP          Universidade C… plan…
#> 10 HUEFS          "Herbario da … UEFS            Universidade E… plan…
#> 11 HUSC           "Herbário Uni… UNISANTA        Universidade S… plan…
#> 12 HVASFw         "Xiloteca do … UNIVASF         Universidade F… plan…
#> 13 JOIw           "Xiloteca Joi… UNIVILLE        Universidade d… plan…
#> 14 MFS            "Herbário Pro… UEPA            Universidade d… plan…
#> 15 NIT            "Herbário de … UFF             Universidade F… plan…
#> 16 SLUI           "Herbário Ros… UEMA            Universidade E… plan…
#> 17 SPFw           "Xiloteca do … USP             Universidade d… plan…
#> # … with 4 more variables: numberOfRecords <chr>, location <chr>,
#> #   keywords_pt <chr>, keywords_en <chr>