Read speciesLink datasets

This function returns the search results from speciesLink network records.

splinkr_records(
  scientificName = NULL,
  barcode = NULL,
  catalogNumber = NULL,
  basisOfRecord = NULL,
  collectionCode = NULL,
  collector = NULL,
  collectorNumber = NULL,
  yearCollected = NULL,
  identifiedBy = NULL,
  yearIdentified = NULL,
  kingdom = NULL,
  phylum = NULL,
  class = NULL,
  order = NULL,
  family = NULL,
  typus = NULL,
  country = NULL,
  stateProvince = NULL,
  county = NULL,
  locality = NULL,
  redlist = NULL,
  maxrecords = NULL
)

Arguments

scientificName
Character vector containing one or more single or compound names, without authors

barcode
Character vector containing one or more institution barcode

catalogNumber
Character vector containing one or more institution catolog number

basisOfRecord
PreservedSpecimen, LivingSpecimen, FossilSpecimen, HumanObservation, MachineObservation, MaterialSample

collectionCode
Character vector containing collections or subcollections acronym

collector
Character vector containing one or more colletor name

collectorNumber
Character vector containing collector number

yearCollected
Numeric four-digits collect year

identifiedBy
Character vector identificator name

yearIdentified
Numeric four-digits year of identifications

kingdom
Character vector containing kingdon name

phylum
Character vector containing phyllum name

class
Character vector containing class name

order
Character vector containing order name

family
Character vector containing family name

typus
If "yes" search only in type material variable

country
Character vector containing collect country

stateProvince
Character vector containing state

county
Character vector containing collect county or city

locality
Character vector containingcollect locality

redlist
If "yes" return only taxa in Portaria MMA 443/2014

maxrecords
Character vector to restrict number of records to return

Value

data.frame() with the records

Examples

if (FALSE) {
# Searching taxon
splinkr_records(scientificName = "Dyckia encholirioides var. rubra")

# Searching taxa collected by collector...
splinkr_records(scientificName = c("Anathallis kleinii",
                                   "Anathallis microphyta"),
                collector = "Siqueira")

# Searching all type specimens from
# one collection
splinkr_records(collectionCode = "FLOR", typus = "yes")

# two collections
splinkr_records(collectionCode = c("FLOR", "FURB"), typus = "yes")
}